+++ docs-src/pages/changelog.svelte
@@ -1,13 +1,13 @@
 <h1>Changelog</h1>
 <h2>v9.1.0 <em>(2023-09-?)</em></h2>
 <ul>
-<li>New components <code>InputRating</code>, <code>Tag</code>.</li>
-<li>Add <code>hideTip</code> property to <code>Popover</code>.</li>
+<li>New components: <code>InputRating</code>, <code>Tag</code>, <code>InputTag</code>.</li>
+<li>Add <code>hideTip</code> and more, to <code>Popover</code>.</li>
 <li>Small bugfixes and improvements.</li>
 </ul>
 <h2>v9.0.5 <em>(2023-09-22)</em></h2>
 <ul>
-<li>Reduce <code>Dialog</code> <code>z-index</code> so that the popups from the dialog show up on top of it.</li>
+<li>Reduce <code>Dialog&#39;s</code> <code>z-index</code> so that the popups from the dialog show up on top of it.</li>
 </ul>
 <h2>v9.0.4, v9.0.3, v9.0.2, v9.0.1 <em>(2023-09-16)</em></h2>
 <ul>